Customer/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
PyMySQL==1.0.3
mysql-connector-python==8.0.33
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==20.1.0
